requirements.txt
asgiref==3.11.0
bleach==6.3.0
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
django-summernote==0.8.20.0
django-taggit==6.1.0
djangorestframework==3.16.1
folium==0.20.0
gunicorn==23.0.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
xyzservices==2025.11.0
asgiref==3.11.0
bleach==6.3.0
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
django-summernote==0.8.20.0
django-taggit==6.1.0
djangorestframework==3.16.1
folium==0.20.0
gunicorn==23.0.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
xyzservices==2025.11.0
